requirements.txt
duckdb>=0.10.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-jose[cryptography]>=3.3.0
reportlab>=4.0.0
pandas>=2.0.0
python-dateutil>=2.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

